requirements.txt
groq
google-cloud-texttospeech
gtts
huggingface
huggingface_hub
langchain
python-dotenv